+++ src/src/consts/index.js
@@ -1,3 +1,4 @@
 export default Object.freeze({
-  site_name: 'yaso28 - Web系エンジニア'
+  site_name: 'yaso28 フリーランスプログラマー',
+  me: 'フリーランスプログラマーyaso28'
 })